botserver/requirements.txt
flask
transformers
langchain
langchain-community
python-dotenv